+++ src/core/node.ts
@@ -19,7 +19,7 @@ export class Node {
     private _isAlive = true // optimization: use binary flags for all these switches
     private _isDetaching = false
 
-    readonly middlewares: IMiddleWareHandler[] = []
+    readonly middlewares: { filter: IMiddleWareFilter; handler: IMiddleWareHandler }[] = []
     private readonly snapshotSubscribers: ((snapshot: any) => void)[] = []
     private readonly patchSubscribers: ((patches: IReversibleJsonPatch) => void)[] = []
     private readonly disposers: (() => void)[] = []
@@ -336,9 +336,12 @@ export class Node {
         return this.parent!.isRunningAction()
     }
 
-    addMiddleWare(handler: IMiddleWareHandler) {
-        // TODO: check / warn if not protected?
-        return registerEventHandler(this.middlewares, handler)
+    addMiddleWare(filter: IMiddleWareFilter, handler: IMiddleWareHandler) {
+        const config = { filter, handler }
+        this.middlewares.push(config)
+        return () => {
+            remove(this.middlewares, config)
+        }
     }
 
     getChildNode(subpath: string): Node {
@@ -489,7 +492,7 @@ import {
     stripPatch
 } from "./json-patch"
 import { walk } from "./mst-operations"
-import { IMiddleWareHandler, createActionInvoker } from "./action"
+import { IMiddleWareFilter, IMiddleWareHandler, createActionInvoker } from "./action"
 import {
     addReadOnlyProp,
     addHiddenFinalProp,
@@ -499,6 +502,7 @@ import {
     registerEventHandler,
     identity,
     noop,
-    freeze
+    freeze,
+    remove
 } from "../utils"
 import { IdentifierCache } from "./identifier-cache"